+++ src/components/Research/index.ts
@@ -1,0 +1,2 @@
+export { default as PaperCard } from "./PaperCard";
+export { default as PaperList } from "./PaperList";